Create multistart_options module
